Fix the issue when an MR is not rebased after updating its target branch (#20)
Fix the issue when an MR is not rebased after updating its target branch.

The `test_insert_new_mr` test in the CI has been failing for this reason for a
while. This fixes it by pushing a mandatory rebase action to an MR after an MR
update. This also removes unnecessary checks on the GitLab API response status.