Optimize git push and API requests
When the number of patches in a stack grows beyond just a few, the
time it takes for gerritlab to perform its operations can become very
long (30 seconds or more).

This commit attempts to improve the situation by making the following
changes:

* Use a single git push to push all commits to their respective
  branches.  This is a major time saver.

* Determine whether an MR needs to be updated or created using only
  the information returned by merge_request.get_all_merge_requests().
  Never perform additional MR queries.

This commit removes the old logic that was required to deal with
unexpected GitLab auto-merge behavior (FIXME: reference needed).  That
problem doesn't seem to happen anymore, which is good because now the
code can be much simpler.

Code has been left in place to collect timings of major operations.

Change-Id: Ia09eb5c8ae8d1a49a9af0a4a5d5eb6e00d3a8e45