After pushing, wait for MRs to stabilize
By stabililze we mean that the "changes_count" field of the MR has a
value of "1".  This ensures that by the time the "git lab" command
terminates, MRs should be fully up-to-date.

Note that this approach will have to be reevaluated if #16 is
implemented (which would allow multiple commits to end up in a single
MR).

Also:
 unit_tests/merge_request_test.py:
  MergeRequestTest:
   Removed WAIT_TIME_BEFORE_VALIDATE, etc since create_merge_requests()
   now waits for the desired state before returning.

Change-Id: Id20d9feb27076afe0603bee334badc815e5528a9