DEPRECATE `private_token` in `.gitreview`; use git credential helper
Current status: `private_token` is stored in `.gitreview` or in git
config files as plaintext.

After this patch: storing a `private_token` in `.gitreview` is
DEPRECATED and will raise an error.

And users will be able to store their `private_token` where it's only
accessible using their git-credential helper of choice.

Change-Id: Ib1c30fa80f143c00e06829a64aa5bf69c121c018