Do not call remote.fetch() at start of create_merge_requests
The call to remote.fetch() takes a noticeable amount of time (about a
second for repos that I work with regularly) and makes Gerritlab feel
sluggish.  When a developer is in a develop/push/CI-TEST/fix loop, the
extra seconds add up without adding value.

Leave it to the user to decide when they want to fetch.

Note: The removal of the call to remote.fetch() actually inadvertently
happened in #54.  This commit makes that change official.

Change-Id: Ib0a6715cc0e2be1b9edc098776619758092f022d